Finish unification
